CLI now wired with the "Client API" in new_api.py
A few things that need to be fixed are:
    - The client API doesn't do SSL at the moment, so it isn't really
      secure to run the CLI from another host (even though it *should*
      be *possible* with how it's wired now.).
    - The CLI still doesn't take user/pass
    - It might be useful for the client API settings at some point to be
      configured with user/pass, though at the moment that would be
      insecure since it's accessible by anyone (chunked in with the
      python package during standard distribution of cobbler)
    - The CLI's "check_setup" function is nice, but only works when run
      on localhost at the moment (it will, in fact, error out if it
      isn't). This should be remedied.
    - Probably a bunch of other things...

The goal of the CobblerAPI class (the "Client API") is to allow people
to just cart around one python file and one configuration file
(currently api_settings.py) for all of their cobbler interaction needs.
This removes unnecessary boiler plate and allows for "central" control
of the messy bits involved in setting up a good connection into cobblerd.

The naming convention "stuff" still needs to be fixed.  The goal is to
rename api.py to core.py so-as to make it clearer what it does, then
name new_api.py to api.py because it should henceforth (once it's
properly tested) be the primary way to interact with Cobbler via python.
And remote.py needs to be renamed such that it's clearly cobblerd's
internal server interface (and that nothing outside of cobblerd should
touch it directly).