Made first step in decrudding remote.py
All *specific* methods of the form *_item (like new_distro...) are now
dynamically built at initialization time of the Cobbler XMLRPC
interface.  This will hopefully simplify the process of having the
remote API built mostly dynamically in the future (one can only hope...)
directly from the datastore.  But for now, it reduces the amount of code
in remote.py, hopefully making it cleaner to maintain and extend.

Next step is to generalize to other *specific* method forms... including
those that aren't quite as clean.  This may require some refactoring of
the API itself instead of just the implementation.