Cleaned up remote.py significantly.
Moved the tests which were at the end of remote.py to
tests/remote_py_tests.py ... There should be a good place to put these
at some point.  The thought is that they might be a good base for more
api testing stuff but that they weren't doing anything but clutter up
remote.py...

IMPORTANT: XXX: BLAH!!!:  changed find_items derivatives from singular
to plural... e.g.  find_distro is now find_distros.  I made no sense for
something which returns a list of matches to be named singularly.  This
breaks the previous API, so be aware (this is why I adjusted the
apitests unit tests...).

Added a curry object to utils.py.  It helps with the dynamic method
generation in remote.py and is just useful in general sometimes.