Fleshed out base items, and fixed minor store bugs
The "current" Items likely don't function quite as advertised, next step
is to write a basic test case for adding them to the store (I have
already done that for Distros).  From there the bulk of the Item testing
should be generalizable enough to have a set of tests that can walk over
all items and try to verify their functionality.

Next I also need to implement "signatures" so that we can keep track of
whether or not the items being read off disk (to be implemented in
datastore) match the Item declarations closely enough to "function".