Alright, everything should be happy now
There is now a config.py file which should house all configuration level
information for the object store at least until we build the cobbler
wide configurator (see comment at top of config.py).

This makes it such that __init__.py no longer needs to hold state and,
as such, may be imported however you want ("import __init__ as store"
vs. "import store").  The key property of python's importer to know for
this to make sense is that if you import a module with a different name,
even if it is the same code, then it won't be the same module.